Moviendo archivos y panel de admin a admins
