+++ puewebcamp/admin/login.php
@@ -18,7 +18,7 @@
   <div class="card">
     <div class="card-body login-card-body">
       <p class="login-box-msg">Inicia sesión aquí</p>
-      <form name="login-admin-form" id="login-admin" method="post" action="modelo-admin.php">
+      <form name="login-admin-form" id="login-admin" method="post" action="login-admin.php">
         <div class="input-group mb-3">
           <input type="text" class="form-control" name="usuario" placeholder="Usuario">
           <div class="input-group-append">
