+++ puewebcamp/admin/templates/navegacion.php
@@ -17,7 +17,7 @@
           <img src="img/1.jpg" class="img-circle elevation-2" alt="User Image">
         </div>
         <div class="info">
-          <a href="#" class="d-block">Fernando López</a>
+          <a href="#" class="d-block"><?php echo $_SESSION['nombre']; ?></a>
         </div>
       </div>
 
@@ -143,6 +143,7 @@
             </ul>
           </li>
 
+          <?php if($_SESSION['nivel'] == 1): ?>
           <!--. Administradores -->
           <li class="nav-item has-treeview">
             <a href="#" class="nav-link">
@@ -167,6 +168,7 @@
               </li>
             </ul>
           </li>
+          <?php endif; ?>
 
         <!--. Testimoniales -->
         <li class="nav-item has-treeview">
